+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 ## 1.0.0 - 2023-09-14
 * First release with scopes of the 5 unified APIs (Accounting, POS, eCommerce, Invoicing & Payment) of Chift and the management of consumers, connections & webhooks.
 * Review of tests and CI-pipeline
+
+## 1.0.1 - 2023-09-21
+* Update accounting model
+* add getOutstandings route